Embedded the signature inside the TRS
The TRS data type now caches the Signature, for efficiency.
Currently the Signature is not used very often, but it is expected that in
further applications it will be needed. For instance, when using different
styles to show defined and constructor symbols when printing the rules

darcs-hash:20080626084100-26539-5bcd3648fe1b39e652b846bf35db57b3202b536e.gz