Got a custom made version of AproVe
This version (made specially for Narradar with a custom strategy)
runs in background, and we communicate with it via a socket.

This patch brings support for this, as well as new options
in the command line client.
The web client has not been updated to this yet

darcs-hash:20081114163004-98b34-0e0e8aa778d92684ec6ffa0c6a9f7ca737e1ce08.gz