Don't show slack work in the logs
Ignore-this: c1bb23d810d69ad9e5219d42a35f40f3

Sadly I have to write this patch title.
The goal with this patch was to *remove* the slack work, but I didn't manage so far.
The problem is in the evaluation of Ors.
The extra work is being done in unwrap, so I will need to redefine it with a custom version for ProgressF.

But even if I managed to fix that, we want to do some work especulatively, when trying several Aprove problems in parallel, so the title of this patch still holds.
It would be nice if we could generate the tree, and then launch the first child of every Or leaf, and then launch a 2nd wave of still not solved Or leafs, and so on.
This seems significatively more complicated though

darcs-hash:20081201190319-b62a5-fb21b802639049abc3304700e2038fba5ec44e52.gz