Changes to the proof EDSL
After several tests, I have lost faith in
monadic solvers. I didn't succeed in emulating
lazy evaluation manually. My tests showed
that solving flat-fb with an impure solver,
by means of runProofT, would take 45 seconds,
whereas a pure version of the solver would do it in 13.

This patch implements BFS for pure solvers,
and a simpler, cleaner set of primitives.
Now the only primitive is runProof.
simplyfy and isSuccess are implemented on top of it.
I kept sliceWorkDone because it is still useful for
unsuccesful proofs.

runProof simply translates a Proof to a MonadPlus
monad, non-deterministically returning a succesful
proof tree. It implements BFS by pruning the tree up
to a certain depth, attempting to find a solution, and
in case of failure try deeper. The scheme is fairly
primitive and definitely not optimal, but I couldn't
come up with a better way.