Simpler, faster BFS for Proofs
the idea is to introduce a Stage command for the Proof monad,
staging expensive computations, like calling AproVE.
These are the object of the search.
Then we translate a proof to a MonadLogic instance and interleave
the items at a MPlusPar node, resulting in a BFS.
In fact, we can define any strategy we like using the very
rich MonadLogic interface.