Disable the top-down optimization in SAT RPO.
This optimization conflicts with the memoization
of formulas optimization. When we memoize a formula,
we might not be memoizing what we expect, since it
can have implicit assumptions due to this top-down
opti. thus being a different formula.

E.g. we may think we are memoizing

f(s(s(x))) > f(s(x))

but what we end up memoizing is

f(s(s(x))) > f(s(x)) assuming that pi(s(x)) = x

which corresponds to a different set of constraints