we cannot use the narrowing depgraph processor before transforming to a relative termination problem
This is an implementation related problem.
IF we do this, then when we build the initial goal
problem we do not have the full graph around
and can produce false positives:
 - no gen rules in scope
 - goal-directed usable rules are empty
   (or specifically, equivalent to regular usable rules)

This idea was intended to enable the solving of
Toyama-like problems, where in absence of gen a
star estimation of the dep graph can show that there
is no connection.
But the same effect can probably be obtained by
using a proper Q-restricted rewriting formalization
of the relative termination processor, so no big
deal, eventually we will have it.