Add a missing equation for involved pairs, where we do not receive a DGraph structure.
This can happen when constructing a new Initial Goal problem
Computation of the EDG requires the usable rules,
which in turn require the involved pairs, which
in turn require the initial graph.

This patch ruthlessly breaks the cycle by assuming
all the pairs are involved, thus allowing for the problem
to be constructed.

NOTE: This patch also unnecessarily generalizes some types
as a result of a failed attempt on properly defining involved
pairs. However the generalized types make sense and should not
break anything.